[ALPS04705712] CPUDVFS: Fix uninitialize coverity
[Detail] CPUDVFS: Fix uninitialize coverity

MTK-Commit-Id: 149c5d372378d55197a94be86e65951e558a4c54

CR-Id: ALPS04705712
Feature: DVFS (Dynamic Voltage and Frequency Scaling)

+++ drivers/misc/mediatek/base/power/upower_v2/mtk_unified_power_mt6779.c
@@ -152,7 +152,7 @@ int upower_bank_to_spower_bank(int upower_bank)
 static void upower_scale_l_cap(void)
 {
 	unsigned int ratio;
-	unsigned int temp;
+	unsigned long long temp;
 	unsigned int max_cap = 1024;
 	int i, j;
 	struct upower_tbl *tbl;
